fix: Resolve VHS workflow failure and pytest conftest conflict
Critical infrastructure fixes to unblock CI/CD and test execution:

VHS Workflow Fix:
- Update VHS version from v0.7.2 to v0.10.0
- Add comprehensive error handling with curl -fsSL flags
- Add installation verification and progress logging
- Fix blocking demo generation in GitHub Actions

Test Infrastructure Fix:
- Resolve pytest conftest plugin collision (tests/unit/tui/conftest.py)
- Consolidate all TUI fixtures to main tests/conftest.py
- Fix test collection: 300/550 → 550/550 (100%)
- Fix test execution: 492/550 passing (89.5%)
- Remove problematic pilot_app fixture with undefined parameter

Files changed: 4 files (1 workflow, 1 deleted, 2 modified)
Lines: +152 documentation, +130 fixtures, +27 workflow

Impact:
- Unblocked VHS demo generation in CI/CD
- Enabled full test suite validation (+250 tests)
- 89.5% test pass rate for core functionality
- Remaining failures are Phase 3 unimplemented features (expected)
